I'd believe in god if it'd work.

+++ src/math_lite.h
@@ -1,2 +1,2 @@
 #pragma once
-long double pow_lite(long double base, long double exp);
+long double pow_lite(float base, float exp);
